Fix MPQ archive scanner, add more file paths and simplify code.
I refactored the MPQ archive scanner so it was simpler and more adaptable. In addition, I modified the code so that a "dummy" archive file entry terminated the list, so that file names can be simply added or removed at will. It will now scan through for all file names. Once found and opened it fall through the name list then start extracting the data. If it ends up at the terminator and no archive has been opened it will then abort with a fatal error. I also added blank lines for output readability. :-)